Changed typing to microsoft standard

+++ src/main.c
@@ -74,7 +74,7 @@
 // Macros to enable Touch LED usage
 #define TOUCH_LED_DIRECTION_BASED(n)											setTouchLEDColor(DEVICE(n), direction_sign ? (direction_sign == 1 ? colorGreen : colorRed) : colorYellow);
 #define TOUCH_LED_BLINK(n, color1, color2)								setTouchLEDColor(DEVICE(n), (i % 2) ? color1 : color2);
-#define TOUCH_LED_CYCLE(n, time)													{ static unsigned long rgb_##n; \
+#define TOUCH_LED_CYCLE(n, time)													{ static int rgb_##n; \
 																													if (i >= rgb_##n) { nextColor(DEVICE(n)); rgb_##n = i + time; } }
 
 // Macros to enable ultrasonic sensor usage
